Tests: Split up long-running e2e UI tests
Change-type: patch
