+++ playwright.config.js
@@ -6,19 +6,17 @@ const config = {
 	use: {
 		headless: true,
 		viewport: {
-			width: 1366, height: 687
+			width: 1366,
+			height: 687
 		},
 		ignoreHTTPSErrors: false,
 		baseURL: process.env.UI_HOST || 'http://jel.ly.fish.local',
 		browserName: 'chromium',
 		launchOptions: {
-			args: [
-				'--disable-gpu',
-				'--disable-dev-shm-usage'
-			]
+			args: [ '--disable-gpu', '--disable-dev-shm-usage' ]
 		}
 	},
-	timeout: 10 * 60 * 1000,
+	timeout: 2 * 60 * 1000,
 	retries: 2,
 	workers: 1
 }
